use a separate module, add doc comments
