+++ external-crates/move/crates/move-binary-format/src/compatibility.rs
@@ -5,16 +5,14 @@
 use std::collections::BTreeSet;
 
 use crate::{
+    compatibility_mode::{CompatibilityMode, ExecutionCompatibilityMode},
     errors::{PartialVMError, PartialVMResult},
     file_format::{AbilitySet, DatatypeTyParameter, Visibility},
     file_format_common::VERSION_5,
-    normalized::{Enum, Function, Module, Struct},
+    normalized::Module,
 };
-use move_core_types::identifier::Identifier;
-use move_core_types::language_storage::ModuleId;
-use move_core_types::{
-    account_address::AccountAddress, identifier::IdentStr, vm_status::StatusCode,
-};
+use move_core_types::vm_status::StatusCode;
+
 // ***************************************************************************
 // ******************* IMPORTANT NOTE ON COMPATIBILITY ***********************
 // ***************************************************************************
@@ -69,255 +67,6 @@ impl Default for Compatibility {
             disallow_change_datatype_type_params: true,
             disallow_new_variants: true,
         }
-    }
-}
-
-// TODO: cli compatability mode impl
-pub trait CompatibilityMode: Default {
-    type Error;
-
-    fn module_id_mismatch(
-        &mut self,
-        old_addr: &AccountAddress,
-        old_name: &IdentStr,
-        new_addr: &AccountAddress,
-        new_name: &IdentStr,
-    );
-
-    fn struct_missing(&mut self, name: &Identifier, old_struct: &Struct);
-    fn struct_ability_mismatch(
-        &mut self,
-        name: &Identifier,
-        old_struct: &Struct,
-        new_struct: &Struct,
-    );
-    fn struct_type_param_mismatch(
-        &mut self,
-        name: &Identifier,
-        old_struct: &Struct,
-        new_struct: &Struct,
-    );
-    fn struct_field_mismatch(
-        &mut self,
-        name: &Identifier,
-        old_struct: &Struct,
-        new_struct: &Struct,
-    );
-
-    fn enum_missing(&mut self, name: &Identifier, old_enum: &Enum);
-    fn enum_ability_mismatch(&mut self, name: &Identifier, old_enum: &Enum, new_enum: &Enum);
-    fn enum_type_param_mismatch(&mut self, name: &Identifier, old_enum: &Enum, new_enum: &Enum);
-    fn enum_new_variant(&mut self, name: &Identifier, old_enum: &Enum, new_enum: &Enum);
-    fn enum_variant_missing(&mut self, name: &Identifier, old_enum: &Enum, tag: usize);
-    fn enum_variant_mismatch(
-        &mut self,
-        name: &Identifier,
-        old_enum: &Enum,
-        new_enum: &Enum,
-        tag: usize,
-    );
-
-    fn function_missing_public(&mut self, name: &Identifier, old_func: &Function);
-    fn function_missing_friend(&mut self, name: &Identifier, old_funcs: &Function);
-    fn function_missing_entry(&mut self, name: &Identifier, old_func: &Function);
-    fn function_signature_mismatch(
-        &mut self,
-        name: &Identifier,
-        old_func: &Function,
-        new_func: &Function,
-    );
-    fn function_lost_public_visibility(&mut self, name: &Identifier, old_func: &Function);
-    fn function_lost_friend_visibility(&mut self, name: &Identifier, old_func: &Function);
-    fn function_entry_compatibility(
-        &mut self,
-        name: &Identifier,
-        old_func: &Function,
-        new_func: &Function,
-    );
-
-    fn friend_module_missing(
-        &mut self,
-        old_friend_module_ids: BTreeSet<ModuleId>,
-        new_friend_module_ids: BTreeSet<ModuleId>,
-    );
-
-    fn finish(&self, _: &Compatibility) -> Result<(), Self::Error>;
-}
-
-struct ExecutionCompatibilityMode {
-    datatype_and_function_linking: bool,
-    datatype_layout: bool,
-    friend_linking: bool,
-    entry_linking: bool,
-    no_new_variants: bool,
-}
-
-impl Default for ExecutionCompatibilityMode {
-    fn default() -> Self {
-        Self {
-            datatype_and_function_linking: true,
-            datatype_layout: true,
-            friend_linking: true,
-            entry_linking: true,
-            no_new_variants: true,
-        }
-    }
-}
-
-impl CompatibilityMode for ExecutionCompatibilityMode {
-    type Error = ();
-
-    fn module_id_mismatch(
-        &mut self,
-        _old_addr: &AccountAddress,
-        _old_name: &IdentStr,
-        _new_addr: &AccountAddress,
-        _new_name: &IdentStr,
-    ) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn struct_missing(&mut self, _name: &Identifier, _old_struct: &Struct) {
-        self.datatype_and_function_linking = false;
-        self.datatype_layout = false;
-    }
-
-    fn struct_ability_mismatch(
-        &mut self,
-        _name: &Identifier,
-        _old_struct: &Struct,
-        _new_struct: &Struct,
-    ) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn struct_type_param_mismatch(
-        &mut self,
-        _name: &Identifier,
-        _old_struct: &Struct,
-        _new_struct: &Struct,
-    ) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn struct_field_mismatch(
-        &mut self,
-        _name: &Identifier,
-        _old_struct: &Struct,
-        _new_struct: &Struct,
-    ) {
-        self.datatype_layout = false;
-    }
-
-    fn enum_missing(&mut self, _name: &Identifier, _old_enum: &Enum) {
-        self.datatype_and_function_linking = false;
-        self.datatype_layout = false;
-    }
-
-    fn enum_ability_mismatch(&mut self, _name: &Identifier, _old_enum: &Enum, _new_enum: &Enum) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn enum_type_param_mismatch(&mut self, _name: &Identifier, _old_enum: &Enum, _new_enum: &Enum) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn enum_new_variant(&mut self, _name: &Identifier, _old_enum: &Enum, _new_enum: &Enum) {
-        self.no_new_variants = false;
-    }
-
-    fn enum_variant_missing(&mut self, _name: &Identifier, _old_enum: &Enum, _tag: usize) {
-        self.datatype_layout = false;
-    }
-
-    fn enum_variant_mismatch(
-        &mut self,
-        _name: &Identifier,
-        _old_enum: &Enum,
-        _new_enum: &Enum,
-        _tag: usize,
-    ) {
-        self.datatype_layout = false;
-    }
-
-    fn function_missing_public(&mut self, _name: &Identifier, _old_func: &Function) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn function_missing_friend(&mut self, _name: &Identifier, _old_func: &Function) {
-        self.friend_linking = false;
-    }
-
-    fn function_missing_entry(&mut self, _name: &Identifier, _old_func: &Function) {
-        self.entry_linking = false;
-    }
-
-    fn function_signature_mismatch(
-        &mut self,
-        _name: &Identifier,
-        old_func: &Function,
-        _new_func: &Function,
-    ) {
-        match old_func.visibility {
-            Visibility::Friend => {
-                self.friend_linking = false;
-            }
-            Visibility::Public => {
-                self.datatype_and_function_linking = false;
-            }
-            Visibility::Private => (),
-        }
-
-        if old_func.is_entry {
-            self.entry_linking = false;
-        }
-    }
-
-    fn function_lost_public_visibility(&mut self, _name: &Identifier, _old_func: &Function) {
-        self.datatype_and_function_linking = false;
-    }
-
-    fn function_lost_friend_visibility(&mut self, _name: &Identifier, _old_func: &Function) {
-        self.friend_linking = false;
-    }
-
-    fn function_entry_compatibility(
-        &mut self,
-        _name: &Identifier,
-        _old_func: &Function,
-        _new_func: &Function,
-    ) {
-        self.entry_linking = false;
-    }
-
-    fn friend_module_missing(
-        &mut self,
-        _old_friend_module_ids: BTreeSet<ModuleId>,
-        _new_friend_module_ids: BTreeSet<ModuleId>,
-    ) {
-        self.friend_linking = false;
-    }
-
-    fn finish(&self, compatability: &Compatibility) -> Result<(), ()> {
-        if compatability.check_datatype_and_pub_function_linking
-            && !self.datatype_and_function_linking
-        {
-            return Err(());
-        }
-        if compatability.check_datatype_layout && !self.datatype_layout {
-            return Err(());
-        }
-        if compatability.check_friend_linking && !self.friend_linking {
-            return Err(());
-        }
-        if compatability.check_private_entry_linking && !self.entry_linking {
-            return Err(());
-        }
-        if compatability.disallow_new_variants && !self.no_new_variants {
-            return Err(());
-        }
-
-        Ok(())
     }
 }
